requirements.txt
asgiref==3.5.2
astroid==2.12.10
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
dill==0.3.5.1
Django==4.1.1
django-ckeditor==6.5.1
django-extensions==3.2.1
django-js-asset==2.0.0
executing==1.2.0
flake8==5.0.4
gunicorn==20.1.0
iniconfig==1.1.1
ipython==8.6.0
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.6
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pylint==2.15.3
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.1
tomli==2.0.1
tomlkit==0.11.5
traitlets==5.5.0
tzdata==2022.2
wcwidth==0.2.5
wrapt==1.14.1
